Move elastic provider service & task settings under their own package similar to authorization and completion.
